custom-detectors/word-detector/python/requirements.txt
google-cloud-vision==3.5.0
fastapi==0.103.2
pydantic==2.4.2
requests==2.31.0
uvicorn==0.23.2
# debugpy== 1.8.0
